requirements.txt
click
PyYAML
requests
